backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic==1.13.1
pydantic>=2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0
PyGithub==2.1.1
openai==1.10.0
anthropic==0.34.0
tenacity==8.2.3
pydantic-core>=2.14.6
asyncpg>=0.29.0
aiofiles==23.2.1
